requirements/common-requirements.txt
# standard imports
python-dateutil
requests

# third party imports
dj_database_url
django
django-auto-logout
django-htmx
django-semantic-admin
django-simple-history
django-widget-tweaks
djangorestframework
docutils
nhs-number
psycopg2-binary
pytest-django
pytest-factoryboy
rapidfuzz
whitenoise
django-filter
